ajuste actualizar producto y ppi foto y retablo

+++ Controllers/Inventario.php
@@ -218,9 +218,10 @@
                                     if($_FILES['txtImgFrame']['name'] == ""){
                                         $photoFraming = $request['framing_img'];
                                     }else{
-                                        if($request['framing_img'] != "category.jpg"){
+                                        /*
+                                        if($request['framing_img'] != "category.jpg" && $request['framing_img'] != null ){
                                             deleteFile($request['framing_img']);
-                                        }
+                                        }*/
                                         $imgFraming = $_FILES['txtImgFrame'];
                                         $photoFraming = 'framing_'.bin2hex(random_bytes(6)).'.png';
                                     }
